+++ tests/testthat/test_tasks.R
@@ -3,6 +3,22 @@ test_that("airpassengers task", {
   expect_task(task)
 })
 
+test_that("monash task", {
+  local_mocked_bindings(
+    download_zenodo_record = function(dataset) {
+      file = system.file("extdata", "m3_yearly_dataset.tsf", package = "mlr3forecast")
+      read_tsf(file)
+    },
+    .package = "mlr3forecast"
+  )
+
+  task = tsk("monash", dataset = "m3_yearly")
+  expect_task(task)
+  expect_identical(task$id, "m3_yearly")
+  expect_identical(tsk("monash", dataset = "m3_yearly", id = "custom")$id, "custom")
+  expect_error(tsk("monash", dataset = "bitcoin_with_missing_values"), class = "Mlr3ErrorInput")
+})
+
 test_that("usaccdeaths task", {
   task = tsk("usaccdeaths")
   expect_task(task)